restructure!: Optimize texture space, reducing number of empty cells
Instead of using the naiive implementation of just keeping all textures
the same size, we are now adjusting the rows to be as few as needed. The
columns are staying fixed so that we can still use the same
`TextureAtlasLayout` for all textures.

This works pretty well and comes at zero cost of the game repo (no extra
code required to make this work). Only had to adjust the trickilm
indices which was very easy.

The sprite bundling process is also a lot faster thanks to ImageMagick.

Also removed the material assignment in the `anim_render.py` script. I
plan to handle that differently (just linking to the parent material).